Struct openssh::SessionBuilder

pub struct SessionBuilder { /* private fields */ }

Build a Session with options.

Implementations

impl SessionBuilder

pub fn get_user(&self) -> Option<&str>

Return the user set in builder.

pub fn get_port(&self) -> Option<&str>

Return the port set in builder.

pub fn user(&mut self, user: String) -> &mut Self

Set the ssh user (ssh -l).

Defaults to None.

pub fn port(&mut self, port: u16) -> &mut Self

Set the port to connect on (ssh -p).

Defaults to None.

pub fn keyfile(&mut self, p: impl AsRef<Path>) -> &mut Self

Set the keyfile to use (ssh -i).

Defaults to None.

pub fn known_hosts_check(&mut self, k: KnownHosts) -> &mut Self

See KnownHosts.

Default KnownHosts::Add.

pub fn connect_timeout(&mut self, d: Duration) -> &mut Self

Set the connection timeout (ssh -o ConnectTimeout).

This value is specified in seconds. Any sub-second duration remainder will be ignored. Defaults to None.

pub fn server_alive_interval(&mut self, d: Duration) -> &mut Self

Set the timeout interval after which if no data has been received from the server, ssh will request a response from the server (ssh -o ServerAliveInterval).

This value is specified in seconds. Any sub-second duration remainder will be ignored. Defaults to None.

pub fn control_directory(&mut self, p: impl AsRef<Path>) -> &mut Self

Set the directory in which the temporary directory containing the control socket will be created.

If not set, openssh will try to use $XDG_STATE_HOME, $HOME/.local/state on unix, and fallback to ./ (the current directory) if it failed.

pub fn clean_history_control_directory(&mut self, clean: bool) -> &mut Self

Clean up the temporary directories with the .ssh-connection prefix in directory specified by SessionBuilder::control_directory, created by previous openssh::Session that is not cleaned up for some reasons (e.g. process getting killed, abort on panic, etc)

Use this with caution, do not enable this if you don’t understand what it does,

pub fn control_persist(&mut self, value: ControlPersist) -> &mut Self

Set the ControlPersist option to configure how long the controlling ssh session should stay alive.

Defaults to ControlPersist::Forever.

pub fn config_file(&mut self, p: impl AsRef<Path>) -> &mut Self

Set an alternative per-user configuration file.

By default, ssh uses ~/.ssh/config. This is equivalent to ssh -F <p>.

Defaults to None.

pub fn compression(&mut self, compression: bool) -> &mut Self

Enable or disable compression (including stdin, stdout, stderr, data for forwarded TCP and unix-domain connections, sftp and scp connections).

Note that the ssh server can forcibly disable the compression.

By default, ssh uses configure value set in ~/.ssh/config.

If ~/.ssh/config does not enable compression, then it is disabled by default.

pub fn jump_hosts<T: AsRef<str>>( &mut self, hosts: impl IntoIterator<Item = T> ) -> &mut Self

Specify one or multiple jump hosts.

Connect to the target host by first making a ssh connection to the jump host described by destination and then establishing a TCP forwarding to the ultimate destination from there.

Multiple jump hops may be specified. This is a shortcut to specify a ProxyJump configuration directive.

Note that configuration directives specified by SessionBuilder do not apply to the jump hosts.

Use ~/.ssh/config to specify configuration for jump hosts.

pub fn user_known_hosts_file( &mut self, user_known_hosts_file: impl AsRef<Path> ) -> &mut Self

Specify the path to the known_hosts file.

The path provided may use tilde notation (~) to refer to the user’s home directory.

The default is ~/.ssh/known_hosts and ~/.ssh/known_hosts2.

pub fn ssh_auth_sock(&mut self, ssh_auth_sock: impl AsRef<Path>) -> &mut Self

Specify the path to the ssh-agent.

The path provided may use tilde notation (~) to refer to the user’s home directory.

The default is None.

pub async fn connect<S: AsRef<str>>( &self, destination: S ) -> Result<Session, Error>

Connect to the host at the given host over SSH using process impl, which will spawn a new ssh process for each Child created.

The format of destination is the same as the destination argument to ssh. It may be specified as either [user@]hostname or a URI of the form ssh://[user@]hostname[:port]. A username or port that is specified in the connection string overrides the one set in the builder (but does not change the builder).

If connecting requires interactive authentication based on STDIN (such as reading a password), the connection will fail. Consider setting up keypair-based authentication instead.

pub fn resolve<'a, 'b>( &'a self, destination: &'b str ) -> (Cow<'a, Self>, &'b str)

SessionBuilder support for destination parsing. The format of destination is the same as the destination argument to ssh.

Examples

use openssh::SessionBuilder;
let b = SessionBuilder::default();
let (b, d) = b.resolve("ssh://test-user@127.0.0.1:2222");
assert_eq!(b.get_port().as_deref(), Some("2222"));
assert_eq!(b.get_user().as_deref(), Some("test-user"));
assert_eq!(d, "127.0.0.1");

pub async fn launch_master(&self, destination: &str) -> Result<TempDir, Error>

Create ssh master session and return TempDir which contains the ssh control socket.