Struct openssh::Stdio

pub struct Stdio(/* private fields */);

Describes what to do with a standard I/O stream for a remote child process when passed to the stdin, stdout, and stderr methods of Command.

Implementations

impl Stdio

pub const fn piped() -> Self

A new pipe should be arranged to connect the parent and remote child processes.

pub const fn null() -> Self

This stream will be ignored. This is the equivalent of attaching the stream to /dev/null.

pub const fn inherit() -> Self

The child inherits from the corresponding parent descriptor.

NOTE that the stdio fd must be in blocking mode, otherwise ssh might not flush all output since it considers (EAGAIN/EWOULDBLOCK) as an error

pub unsafe fn from_raw_fd_owned(fd: RawFd) -> Self

Stdio::from_raw_fd_owned takes ownership of the fd passed in and closes the fd on drop.

NOTE that the fd will be put into blocking mode, then it will be closed when Stdio is dropped.

Safety

fd - must be a valid fd and must give its ownership to Stdio.